GLPv3.
Due to people stealing the code & not naming the original authors which is required by the previous license, i decided to take measurements and change the source code to be GPLv3. This prevents people from further stealing our changes we make & us being able to sue them in case they do without acknowledgement of *ALL* contributors.